+++ juniper_graphql_ws/src/graphql_ws/server_message.rs
@@ -1,7 +1,6 @@
-use juniper::{ExecutionError, Value};
 use serde::Serialize;
 
-pub use crate::server_message::ErrorPayload;
+pub use crate::server_message::{ErrorPayload, NextPayload as DataPayload};
 
 /// The payload for errors that are not associated with a GraphQL operation.
 #[derive(Debug, Eq, PartialEq, Serialize)]
@@ -9,20 +8,6 @@ pub use crate::server_message::ErrorPayload;
 pub struct ConnectionErrorPayload {
     /// The error message.
     pub message: String,
-}
-
-/// Sent after execution of an operation. For queries and mutations, this is sent to the client
-/// once. For subscriptions, this is sent for every event in the event stream.
-#[derive(Debug, PartialEq, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct DataPayload<S> {
-    /// The result data.
-    pub data: Value<S>,
-
-    /// The errors that have occurred during execution. Note that parse and validation errors are
-    /// not included here. They are sent via Error messages.
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub errors: Vec<ExecutionError<S>>,
 }
 
 /// ServerMessage defines the message types that servers can send.