+++ juniper_graphql_ws/src/server_message.rs
@@ -3,8 +3,28 @@
 use std::{any::Any, marker::PhantomPinned};
 
 use derive_more::with_trait::Debug;
-use juniper::GraphQLError;
+#[cfg(doc)]
+use juniper::futures::Stream;
+use juniper::{ExecutionError, GraphQLError, Value};
 use serde::{Serialize, Serializer};
+
+/// Payload to be send after execution of an operation.
+///
+/// - For queries and mutations, this is sent to the client once.
+/// - For subscriptions, this is sent for every event in the event [`Stream`].
+#[derive(Debug, PartialEq, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct NextPayload<S> {
+    /// Execution result data.
+    pub data: Value<S>,
+
+    /// Errors that have occurred during execution.
+    ///
+    /// Note, that parse and validation errors are not included here. They are sent via
+    /// [`ErrorPayload`].
+    #[serde(skip_serializing_if = "Vec::is_empty")]
+    pub errors: Vec<ExecutionError<S>>,
+}
 
 /// Payload for errors that can happen before execution.
 ///